Removed component and moved state to artexhibit page
