--- src/components/ArtCollection.css
@@ -1,6 +1,0 @@
-.art-collection {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 20px;
-  padding: 20px;
-}
